Add a makefile for docs generation
I'm eventually going to move most of the tasks managed by dev.yml into a
makefile. This takes the dev command a bit further by automatically
running when the source files have changed.

Doing this so that folks outside of Shopify can use a common tool to run
our commands if they want to.

This actually fixes a bug in that its not just sibling_groups that needs
to be checked, attributes.yml also needs to be there.

This way make will just "do the right thing" assuming bin/generate_docs
works correctly