Update the cue schema for full validation and referential integrity
There's a lot more that _can_ be done here, but this is probably enough.

This does a few things:
- Ensures attribute names are unique
- Ensures attribute values are unique per attribute
- Does the same as ^ for categories
- Ensures GID's are unique, and for the cases where the GID's are used
  as references, ensure the name matches the canonical source being
  pointed to.

Someone else could take this further by ensuring that the regex is
recursively correct, but I don't think that's worth the effort.

This schema expects some other process to import categories.json and
attributes.json into the schema folder. That will be done via make in
the next commit